Module Tjr_monad

Generic monads

Monad types

include Tjr_monad__.Monad_intf

Monad ops

type ('a, 't) m

type 't monad_ops = {

return : a. 'a -> ('a, 't) m;

bind : a b. ('a, 't) m -> ('a -> ('b, 't) m) -> ('b, 't) m;

}

type 't async = (unit -> (unit, 't) m) -> (unit, 't) m

The async operation completes with unit almost immediately; the argument may be a long-running computation; it is scheduled for execution.

NOTE Because even creating an 'a m in lwt schedules computation, we shield the argument to async to avoid computation.

type ('s, 't) with_state = {

with_state : a. (state:'s -> set_state:('s -> (unit, 't) m) -> ('a, 't) m) -> ('a, 't) m;

}

A type to allow parameterization over a generic notion of "part of state". Presumably the part of the state is locked while the monadic operation completes, and then the lock is released.

NOTE see also with_imperative_ref in Tjr_monad

Events, with signal (similar to promise)

type 'a event

The type of events that resolve to 'a

type 't event_ops = {

ev_create : a. unit -> ('a event, 't) m;

ev_wait : a. 'a event -> ('a, 't) m;

ev_signal : a. 'a event -> 'a -> (unit, 't) m;

}

Multiple threads can be waiting on an event. Only one thread should signal the event. It is an error to signal the same event more than once. If this happens, the new signalled value is discarded and no further actions are taken. FIXME perhaps we should throw an exception or something. An event is a bit like a channel that multiple readers can register on (and all will receive any msgs), and which only carries at most one msg.

NOTE lwt has two types, 'a t for promises and 'a u for resolvers. These are implemented using the same underlying structure.

Mutexes and cvars

type ('mutex, 'cvar, 't) mutex_ops = {

create_mutex : unit -> ('mutex, 't) m;

create_cvar : unit -> ('cvar, 't) m;

lock : 'mutex -> (unit, 't) m;

signal : 'cvar -> (unit, 't) m;

unlock : 'mutex -> (unit, 't) m;

wait : 'mutex -> 'cvar -> (unit, 't) m;

}

What we need from mutexes and condition vars

State passing

type 'a state_passing = 'a State_passing.state_passing

val sp_to_fun : ('a, 'b State_passing.state_passing) Tjr_monad__.Monad_intf.m -> 'b -> 'a * 'b

val sp_of_fun : ('a -> 'b * 'a) -> ('b, 'a State_passing.state_passing) Tjr_monad__.Monad_intf.m

module State_passing : sig ... end

A generic state-passing monad.

val state_passing_monad_ops : unit -> 'a State_passing.state_passing Tjr_monad__.Monad_intf.monad_ops

Imperative

module Imperative : sig ... end

A trivial monad, where bind is just function composition, and return is the identity. Used when performing imperative operations eg Unix.read/write. AKA the identity monad

type imperative = Imperative.imperative

val imperative_monad_ops : Imperative.imperative Tjr_monad__.Monad_intf.monad_ops

With-state

type ('s, 't) with_state = ('s, 't) Tjr_monad__.Monad_intf.with_state = {

with_state : a. (state:'s -> set_state:('s -> (unit, 't) m) -> ('a, 't) m) -> ('a, 't) m;

}

type ('s,'t) with_state = ('s,'t) Monad_intf.with_state = {
  with_state: 
    'a. 
      (state:'s -> 
       set_state:('s -> (unit,'t)m) -> 
       ('a,'t) m)
    -> ('a,'t)m
}

Iteration

val iter_m : monad_ops:'a monad_ops -> ('b -> ('b option, 'a) m) -> 'b -> ('b, 'a) m

val join_seq : monad_ops:'t monad_ops -> (unit, 't) m list -> (unit, 't) m

A generic version of join, which waits for each m to resolve FIXME deprecated

With_lwt

module With_lwt : sig ... end

type lwt = With_lwt.lwt

val lwt_monad_ops : With_lwt.lwt Tjr_monad__.Monad_intf.monad_ops

Util

val with_imperative_ref : monad_ops:'a monad_ops -> 'b Stdlib.ref -> ('b, 'a) with_state