Module Tjr_monad__.Event

type 'a event

The type of events that resolve to 'a

type 't event_ops = {

ev_create : a. unit -> ('a event, 't) Tjr_monad__.Monad_ops.m;

ev_wait : a. 'a event -> ('a, 't) Tjr_monad__.Monad_ops.m;

ev_signal : a. 'a event -> 'a -> (unit, 't) Tjr_monad__.Monad_ops.m;

}

Multiple threads can be waiting on an event. Only one thread should signal the event. It is an error to signal the same event more than once. If this happens, the new signalled value is discarded and no further actions are taken. FIXME perhaps we should throw an exception or something. An event is a bit like a channel that multiple readers can register on (and all will receive any msgs), and which only carries at most one msg.

NOTE lwt has two types, 'a t for promises and 'a u for resolvers. These are implemented using the same underlying structure.