Module Tjr_monad__With_lwt

Lwt basics

type lwt

val lwt_monad_ops : lwt Tjr_monad__.Monad_intf.monad_ops

val return : 'a -> ('a, lwt) Tjr_monad__.Monad_intf.m

val (>>=) : ('a, lwt) Tjr_monad__.Monad_intf.m -> ('a -> ('b, lwt) Tjr_monad__.Monad_intf.m) -> ('b, lwt) Tjr_monad__.Monad_intf.m

val monad_ops : lwt Tjr_monad__.Monad_intf.monad_ops

val to_lwt : ('a, lwt) Tjr_monad__.Monad_intf.m -> 'a Lwt.t

Conversions between lwt and our parameterized monad

val from_lwt : 'a Lwt.t -> ('a, lwt) Tjr_monad__.Monad_intf.m

Lwt events

module Lwt_event_ops : sig ... end

val lwt_event_ops : lwt Tjr_monad__.Monad_intf.event_ops

val event_ops : lwt Tjr_monad__.Monad_intf.event_ops

Lwt mutex ops

module Lwt_mutex_ops : sig ... end

val lwt_mutex_ops : (Lwt_mutex.t, unit Lwt_condition.t, lwt) Tjr_monad__.Monad_intf.mutex_ops

val mutex_ops : (Lwt_mutex.t, unit Lwt_condition.t, lwt) Tjr_monad__.Monad_intf.mutex_ops

Opening files (convenience)

module Lwt_file_ops : sig ... end

val lwt_file_ops : < close : Lwt_unix.file_descr -> (unit, lwt) Tjr_monad__.Monad_intf.m; open_file : fn:string -> create:bool -> trunc:bool -> (Lwt_unix.file_descr, lwt) Tjr_monad__.Monad_intf.m; sync : Lwt_unix.file_descr -> (unit, lwt) Tjr_monad__.Monad_intf.m; >

val file_ops : < close : Lwt_unix.file_descr -> (unit, lwt) Tjr_monad__.Monad_intf.m; open_file : fn:string -> create:bool -> trunc:bool -> (Lwt_unix.file_descr, lwt) Tjr_monad__.Monad_intf.m; sync : Lwt_unix.file_descr -> (unit, lwt) Tjr_monad__.Monad_intf.m; >

Other useful defns

val with_ref : 'a Stdlib.ref -> ('a, lwt) Tjr_monad__.Monad_intf.with_state

val with_locked_ref : 'a Stdlib.ref -> ('a, lwt) Tjr_monad__.Monad_intf.with_state

FIXME really, we should always use a lock, unless we know that only a single thread accesses the state

val async : lwt Tjr_monad__.Monad_intf.async

async for lwt

val yield : unit -> unit Lwt.t

val sleep : float -> unit Lwt.t